+++ generators/python/core_utilities/sdk/http_client.py
@@ -1,23 +1,23 @@
-import http
-import random
+import asyncio
+from contextlib import _AsyncGeneratorContextManager, _GeneratorContextManager
+from random import random
 import time
 import httpx
 import typing
 from functools import wraps
-from .request_options import RequestOptions
 import re
-import email
+import email.utils
 
 INITIAL_RETRY_DELAY_SECONDS = .5
 MAX_RETRY_DELAY_SECONDS = 10
 MAX_RETRY_DELAY_SECONDS_FROM_HEADER = 30
 
 
-# Inspired by the urllib3 implementation.
-def _parse_retry_after(response_headers: httpx.Headers) -> typing.Optional[int]:
+def _parse_retry_after(response_headers: httpx.Headers) -> typing.Optional[float]:
     """
-    Determine how long to wait before retrying a request.
     This function parses the `Retry-After` header in a HTTP response and returns the number of seconds to wait.
+
+    Inspired by the urllib3 retry implementation.
     """
     retry_after_ms = response_headers.get("retry-after-ms")
     if retry_after_ms is not None:
@@ -32,7 +32,7 @@ def _parse_retry_after(response_headers: httpx.Headers) -> typing.Optional[int]:
 
     # Attempt to parse the header as an int.
     if re.match(r"^\s*[0-9]+\s*$", retry_after):
-        seconds = int(retry_after)
+        seconds = float(retry_after)
     # Fallback to parsing it as a date.
     else:
         retry_date_tuple = email.utils.parsedate_tz(retry_after)
@@ -53,7 +53,13 @@ def _parse_retry_after(response_headers: httpx.Headers) -> typing.Optional[int]:
 
     return seconds
 
-def _retry_timeout(response: httpx.Response, max_retries: int, retries: int) -> bool:
+def _retry_timeout(response: httpx.Response, retries: int) -> float:
+    """
+    Determine the amount of time to wait before retrying a request.
+    This function begins by trying to parse a retry-after header from the response, and then proceeds to use exponential backoff
+    with a jitter to determine the number of seconds to wait.
+    """
+
     # If the API asks us to wait a certain amount of time (and it's a reasonable amount), just do what it says.
     retry_after = _parse_retry_after(response.headers)
     if retry_after is not None and retry_after <= MAX_RETRY_DELAY_SECONDS_FROM_HEADER:
@@ -66,6 +72,9 @@ def _retry_timeout(response: httpx.Response, max_retries: int, retries: int) -> 
     timeout = retry_delay * (1 - 0.25 * random())
     return timeout if timeout >= 0 else 0
 
+def _should_retry(response: httpx.Response) -> bool:
+    retriable_400s = [429, 408, 409]
+    return response.status_code >= 500 or response.status_code in retriable_400s
 
 class HttpClient:
     def __init__(self, *, httpx_client: httpx.Client):
@@ -73,19 +82,33 @@ class HttpClient:
 
     # Ensure that the signature of the `request` method is the same as the `httpx.Client.request` method
     @wraps(httpx.Client.request)
-    def request(self, request_options: typing.Optional[RequestOptions] = None, retries: typing.Optional[int] = 0, *args, **kwargs):
-        try:
-            self.httpx_client.request(*args, **kwargs)
-        except Exception as e:
-            if (request_options is not None and request_options.get("max_retries") is not None and request_options.get("max_retries") > retries):
-                time.sleep(_retry_timeout)
-                self.request(request_options, retries + 1, *args, **kwargs)
-            else:
-                raise e
-
+    def request(self, *args: typing.Any, max_retries: int = 0, retries: int = 0, **kwargs: typing.Any) -> httpx.Response:
+        response = self.httpx_client.request(*args, **kwargs)
+        if _should_retry(response=response):
+            if (max_retries > retries):
+                time.sleep(_retry_timeout(response=response, retries=retries))
+                return self.request(max_retries=max_retries, retries=retries + 1, *args, **kwargs)
+        return response
+    
+    @wraps(httpx.Client.stream)
+    def stream(self, *args: typing.Any, max_retries: int = 0, retries: int = 0, **kwargs: typing.Any) -> _GeneratorContextManager[httpx.Response]:
+        return self.httpx_client.stream(*args, **kwargs)
         
-
 
 class AsyncHttpClient:
     def __init__(self, *, httpx_client: httpx.AsyncClient):
         self.httpx_client = httpx_client
+
+    # Ensure that the signature of the `request` method is the same as the `httpx.Client.request` method
+    @wraps(httpx.AsyncClient.request)
+    async def request(self, *args: typing.Any, max_retries: int = 0, retries: int = 0, **kwargs: typing.Any) -> httpx.Response:
+        response = await self.httpx_client.request(*args, **kwargs)
+        if _should_retry(response=response):
+            if (max_retries > retries):
+                await asyncio.sleep(_retry_timeout(response=response, retries=retries))
+                return await self.request(max_retries=max_retries, retries=retries + 1, *args, **kwargs)
+        return response
+    
+    @wraps(httpx.AsyncClient.request)
+    async def stream(self, *args: typing.Any, max_retries: int = 0, retries: int = 0, **kwargs: typing.Any) -> _AsyncGeneratorContextManager[httpx.Response]:
+        return self.httpx_client.stream(*args, **kwargs)