+++ seed/python-sdk/auth-environment-variables/src/seed/core/client_wrapper.py
@@ -4,6 +4,8 @@ import typing
 
 import httpx
 
+from .http_client import AsyncHttpClient, HttpClient
+
 
 class BaseClientWrapper:
     def __init__(self, *, x_another_header: str, api_key: str, base_url: str):
@@ -28,10 +30,10 @@ class BaseClientWrapper:
 class SyncClientWrapper(BaseClientWrapper):
     def __init__(self, *, x_another_header: str, api_key: str, base_url: str, httpx_client: httpx.Client):
         super().__init__(x_another_header=x_another_header, api_key=api_key, base_url=base_url)
-        self.httpx_client = httpx_client
+        self.httpx_client = HttpClient(httpx_client=httpx_client)
 
 
 class AsyncClientWrapper(BaseClientWrapper):
     def __init__(self, *, x_another_header: str, api_key: str, base_url: str, httpx_client: httpx.AsyncClient):
         super().__init__(x_another_header=x_another_header, api_key=api_key, base_url=base_url)
-        self.httpx_client = httpx_client
+        self.httpx_client = AsyncHttpClient(httpx_client=httpx_client)