+++ seed/python-sdk/exhaustive/infinite-timeout/src/seed/resources/endpoints/resources/union/client.py
@@ -56,6 +56,8 @@ class UnionClient:
             timeout=request_options.get("timeout_in_seconds")
             if request_options is not None and request_options.get("timeout_in_seconds") is not None
             else None,
+            retries=0,
+            max_retries=request_options.get("max_retries") if request_options is not None else 0,  # type: ignore
         )
         if 200 <= _response.status_code < 300:
             return pydantic.parse_obj_as(Animal, _response.json())  # type: ignore
@@ -102,6 +104,8 @@ class AsyncUnionClient:
             timeout=request_options.get("timeout_in_seconds")
             if request_options is not None and request_options.get("timeout_in_seconds") is not None
             else None,
+            retries=0,
+            max_retries=request_options.get("max_retries") if request_options is not None else 0,  # type: ignore
         )
         if 200 <= _response.status_code < 300:
             return pydantic.parse_obj_as(Animal, _response.json())  # type: ignore