+++ seed/python-sdk/exhaustive/infinite-timeout/src/seed/resources/inlined_requests/client.py
@@ -69,6 +69,8 @@ class InlinedRequestsClient:
             timeout=request_options.get("timeout_in_seconds")
             if request_options is not None and request_options.get("timeout_in_seconds") is not None
             else None,
+            retries=0,
+            max_retries=request_options.get("max_retries") if request_options is not None else 0,  # type: ignore
         )
         if 200 <= _response.status_code < 300:
             return pydantic.parse_obj_as(ObjectWithOptionalField, _response.json())  # type: ignore
@@ -128,6 +130,8 @@ class AsyncInlinedRequestsClient:
             timeout=request_options.get("timeout_in_seconds")
             if request_options is not None and request_options.get("timeout_in_seconds") is not None
             else None,
+            retries=0,
+            max_retries=request_options.get("max_retries") if request_options is not None else 0,  # type: ignore
         )
         if 200 <= _response.status_code < 300:
             return pydantic.parse_obj_as(ObjectWithOptionalField, _response.json())  # type: ignore