+++ seed/python-sdk/no-environment/src/seed/core/client_wrapper.py
@@ -4,6 +4,8 @@ import typing
 
 import httpx
 
+from .http_client import AsyncHttpClient, HttpClient
+
 
 class BaseClientWrapper:
     def __init__(self, *, token: typing.Union[str, typing.Callable[[], str]], base_url: str):
@@ -34,7 +36,7 @@ class SyncClientWrapper(BaseClientWrapper):
         self, *, token: typing.Union[str, typing.Callable[[], str]], base_url: str, httpx_client: httpx.Client
     ):
         super().__init__(token=token, base_url=base_url)
-        self.httpx_client = httpx_client
+        self.httpx_client = HttpClient(httpx_client=httpx_client)
 
 
 class AsyncClientWrapper(BaseClientWrapper):
@@ -42,4 +44,4 @@ class AsyncClientWrapper(BaseClientWrapper):
         self, *, token: typing.Union[str, typing.Callable[[], str]], base_url: str, httpx_client: httpx.AsyncClient
     ):
         super().__init__(token=token, base_url=base_url)
-        self.httpx_client = httpx_client
+        self.httpx_client = AsyncHttpClient(httpx_client=httpx_client)