+++ generators/python/sdk/CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.8] - 2024-03-02
+
+- Improvement: Introduces a `max_retries` parameter to the RequestOptions dict accepted by all requests. This parameter will retry requests automatically, with exponential backoff and a jitter. The client will automatically retry requests of a 5XX status code, or certain 4XX codes (429, 408, 409).
+
+  ```python
+    client\
+      .imdb\
+      .create_movie(
+        request=CreateMovieRequest(title="title", rating=4.3),
+        request_options={
+          "max_retries": 5,
+        }
+      )
+  ```
+
 ## [0.11.7] - 2024-02-27
 
 - Improvement: Introduces a flag `use_str_enums` to swap from using proper Enum classes to using Literals to represent enums. This change allows for forward compatibility of enums, since the user will receive the string back.
@@ -21,10 +36,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Operand = typing.Literal[">", "=", "less_than"]
   ```
 
-
 ## [0.11.6] - 2024-02-26
 
 - Improvement: You can now specify envvars to scan for headers, not just auth scheme headers.
+
   ```
   # OpenAPI
   x-fern-global-headers:
@@ -33,7 +48,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
      optional: true
      env: MY_API_KEY
   ```
+
   ... or ...
+
   ```
   # Fern Definition
   getAllUsers:
@@ -45,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         X-API-KEY: string
         env: MY_API_KEY
   ```
+
   the generated client will look like
 
   ```python